requirements.txt
crewai
python-dotenv
pandas
numpy
pymongo
matplotlib
seaborn
python-docx
reportlab
openai
faiss-cpu
streamlit
transformers
peft
accelerate
torch
tqdm
datasets